Comments/requirements.txt
anyio==3.5.0
asgiref==3.4.1
backports.cached-property==1.0.1
click==8.0.3
colorama==0.4.4
fastapi==0.72.0
graphql-core==3.1.7
h11==0.13.0
httptools==0.3.0
idna==3.3
pydantic==1.9.0
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
sentinel==0.3.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
strawberry-graphql==0.94.0
typing-extensions==4.0.1
uvicorn==0.17.0
watchgod==0.7
websockets==10.1
